temrinal: show compute server bar at top of terminal; also show exit message when switching between compute servers
